Bug 14358 Make changes to DominantResourceCalculator to make use of Disk
* Good and easy to read article on Max-Min fairness and weighted max-min fairness.
http://www.mathcs.emory.edu/~cheung/Courses/558/Syllabus/11-Fairness/Fair.html
* Dominant Resource Fairness (DRF) is a generalization of max-min fairness to multiple resource types.
* Let me use an example from the paper to explain DRF
Consider a system with of 9 CPUs, 18 GB RAM, and two users, where user A runs tasks with demand vector <1CPU, 4 GB>, and user B runs tasks with demand vector <3 CPUs, 1 GB> each.
In the above scenario, each task from user A consumes 1/9 of the total CPUs and 2/9 of the total memory, so user A’s dominant resource is memory.
Each task from user B consumes 1/3 of the total CPUs and 1/18 of the total memory, so user B s dominant resource is CPU.
DRF will equalize users’ dominant shares, giving the following allocation
Three tasks for user A, with a total of  <3 CPUs, 12 GB> , and two tasks for user B , with a total of <6 CPUs, 2 GB>.
With this allocation, each user ends up with the same dominant share, i.e., user A gets 2/3 of RAM, while user B gets 2/3 of the CPUs.

* Implemented a resource calculator derived from DominantResourceCalculator to implement this behaviour.
* Took care to ensure we do not have incorrect behaviour when disk is not specified (disk=0.0) in the Resource object.

* Changed scheduler to capacity scheduler in yarn-site.xml
* Changed yarn.scheduler.capacity.resource-calculator to org.apache.hadoop.yarn.util.resource.DiskBasedDominantResourceCalculator
  in /opt/mapr/hadoop/hadoop-2.4.1/etc/hadoop/capacity-scheduler.xml
* Was able to run Teragen and Terasort successfully.

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-common/src/main/java/org/apache/hadoop/yarn/util/resource/Resources.java
@@ -177,7 +177,7 @@ public class Resources {
   public static Resource multiplyTo(Resource lhs, double by) {
     lhs.setMemory((int)(lhs.getMemory() * by));
     lhs.setVirtualCores((int)(lhs.getVirtualCores() * by));
-    lhs.setDisks((int)(lhs.getDisks() * by));
+    lhs.setDisks(lhs.getDisks() * by);
     return lhs;
   }
